pyproject.toml
[tool.poetry]
name = "caminho-dos-pes"
version = "0.1.0"
description = "Sistema para clinica de podologia"
authors = ["paulovitorgm <paulovitorgaspmelo@hotmail.com>"]
license = "MIT"
readme = "README.md"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"


[tool.poetry.dependencies]
python = "^3.12"
django = "^5.1.2"
psycopg = "^3.2.3"
djangorestframework = "^3.15.2"
django-filter = "^24.3"
markdown = "^3.7"
django-extensions = "^3.2.3"
coverage = "^7.6.4"



[tool.poetry.group.dev.dependencies]
taskipy = "^1.14.0"
ruff = "^0.7.0"
faker = "^30.8.0"

[tool.poetry.group.dev]
optional = true


[tool.ruff]
line-length = 85
exclude = ['settings.py']
extend-exclude = ['migrations', 'manage.py']



[tool.ruff.format]
preview = true
quote-style = 'single'


[tool.ruff.lint]
preview = true
select = ['I', 'F', 'E', 'W', 'PL', 'PT']
ignore = ['PT009']


[tool.taskipy.tasks]
lint = 'ruff check .; ruff check . --diff'
format = 'ruff check . --fix && ruff format .'
run = 'python manage.py runserver'
m = 'python manage.py'
pre_test = 'task lint'
test = 'coverage run --source="." manage.py test'
post_test = 'coverage html'

